+++ src/pyrepo/commands/mkgithub.py
@@ -1,6 +1,9 @@
+from base64 import b64encode
 import logging
 import re
 import click
+from nacl import encoding, public
+from ..clack import ConfigurableCommand
 from ..gh import GitHub
 from ..project import Project, with_project
 from ..util import cpe_no_tb
@@ -8,12 +11,28 @@ from ..util import cpe_no_tb
 log = logging.getLogger(__name__)
 
 
-@click.command()
+@click.command(cls=ConfigurableCommand, allow_config=["codecov_token"])
+@click.option(
+    "--codecov-token",
+    help="Value for CODECOV_TOKEN actions secret",
+    metavar="SECRET",
+    envvar="CODECOV_TOKEN",
+    show_envvar=True,
+)
+@click.option(
+    "--no-codecov-token", is_flag=True, help="Do not set CODECOV_TOKEN actions secret"
+)
 @click.option("-P", "--private", is_flag=True, help="Make the new repo private")
 @click.option("--repo-name", metavar="NAME", help="Set the name of the repository")
 @with_project
 @cpe_no_tb
-def cli(project: Project, repo_name: str | None, private: bool) -> None:
+def cli(
+    project: Project,
+    repo_name: str | None,
+    private: bool,
+    codecov_token: str | None,
+    no_codecov_token: bool,
+) -> None:
     """Create a repository on GitHub for the local project and upload it"""
     if repo_name is None:
         repo_name = project.details.repo_name
@@ -61,9 +80,33 @@ def cli(project: Project, repo_name: str | None, private: bool) -> None:
                     "description": "Update a Python dependency",
                 }
             )
+            if not no_codecov_token:
+                if codecov_token:
+                    for scope in ["actions", "dependabot"]:
+                        log.info("Setting CODECOV_TOKEN secret (%s)", scope)
+                        secrets = ghrepo / scope / "secrets"
+                        pubkey = (secrets / "public-key").get()
+                        (secrets / "CODECOV_TOKEN").put(
+                            {
+                                "encrypted_value": encrypt_secret(
+                                    pubkey["key"], codecov_token
+                                ),
+                                "key_id": pubkey["key_id"],
+                            }
+                        )
+                else:
+                    log.warn("CODECOV_TOKEN value not set; not setting secret")
     log.info('Setting "origin" remote')
     if "origin" in project.repo.get_remotes():
         project.repo.rm_remote("origin")
     project.repo.add_remote("origin", r["ssh_url"])
     log.info("Pushing to origin")
     project.repo.run("push", "-u", "origin", "refs/heads/*", "refs/tags/*")
+
+
+def encrypt_secret(public_key: str, secret_value: str) -> str:
+    """Encrypt a string for use as a GitHub secret using a public key"""
+    pkey = public.PublicKey(public_key.encode("utf-8"), encoding.Base64Encoder)
+    sealed_box = public.SealedBox(pkey)
+    encrypted = sealed_box.encrypt(secret_value.encode("utf-8"))
+    return b64encode(encrypted).decode("utf-8")